week08_MCP/mtr-mcp-example/requirements.txt
fastmcp
mcp
requests
langgraph
langchain-aws
langchain-core
boto3
python-dotenv
